test-requirements.txt
mock>=1.0
nose>=1.3
coverage
testtools>=0.9.36,!=1.2.0
tox
nose-cov
pylint
flake8
requests-mock
PyYAML==5.1.1
# for blueprints validate
nose-exclude

# for integration tests
git+https://github.com/cloudify-cosmo/cloudify-common@5.1.0#egg=cloudify-common[dispatcher]==5.1.0
git+https://github.com/cloudify-cosmo/cloudify-cli@5.1.0#egg=cloudify-cli==5.1.0
pytest==4.6.3
